worker/requirements.txt
kafka-python==1.4.7
requests
python-dotenv
httpx
schedule
apscheduler
asyncio
confluent_kafka
plotly
aiohttp
redis-om
websockets
aiokafka
pandas
uvicorn
fastapi
SQLAlchemy
redis
aiohttp
aiokafka
aiomysql
aioredis